fix for the tagging descriptor. The deletion method for descriptors is __delete__ not __del__: http://docs.python.org/ref/descriptors.html
The __del__ method is for the base descriptor instance itself.
This was causing problems in apache and mod_wsgi at shutdown.
This was not noticed until shutdown because of the registry which holds onto 
the models until the process terminates and the cyclical garbage collection
kicks in.

If you were seeing errors like:
[Fri Aug 15 01:33:39 2008] [error] Exception exceptions.TypeError: '__del__() takes exactly 2 arguments (1 given)' in <bound method TagDescriptor.__del__ of <tagging.managers.TagDescriptor object at 0x1a7e5d0>> ignored
[Fri Aug 15 01:33:39 2008] [error] Exception exceptions.TypeError: '__del__() takes exactly 2 arguments (1 given)' in <bound method TagDescriptor.__del__ of <tagging.managers.TagDescriptor object at 0x1a7e5d0>> ignored

This bug was the reason for it.



git-svn-id: http://django-tagging.googlecode.com/svn@143 83e7428b-ec2a-0410-86f2-bf466d0e5e72